requirements.txt
absl-py==0.9.0
albumentations==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.1
cachetools==4.1.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==2.2.1
comm==0.1.4
cryptography==3.2
cycler==0.10.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.5.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
efficientnet-pytorch==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.12.2
flatbuffers==2.0.7
fsspec==2023.1.0
future==0.18.2
gast==0.3.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
huggingface-hub==0.16.4
hyperopt==0.2.7
idna==2.10
imageio==2.31.2
importlib-metadata==1.7.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libclang==16.0.6
Markdown==3.2.2
MarkupSafe==2.1.3
matplotlib==3.2.2
matplotlib-inline==0.1.6
mistune==3.0.2
munch==4.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
networkx==2.6.3
nltk==3.5
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.20.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
opencv-python-headless==4.8.1.78
opt-einsum==3.2.1
packaging==23.2
pandas==1.0.5
pandocfilters==1.5.0
paramiko==2.7.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pretrainedmodels==0.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==3.12.2
psutil==5.9.7
ptyprocess==0.7.0
py4j==0.10.9.7
pyaml==23.5.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.17.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyWavelets==1.3.0
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
qudida==0.0.4
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==0.23.1
scikit-optimize==0.9.0
scipy==1.4.1
seaborn==0.10.1
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
six==1.15.0
sniffio==1.3.0
soupsieve==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.7.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.17.1
texttable==1.6.2
threadpoolctl==2.1.0
tifffile==2021.11.2
timm==0.9.2
tinycss2==1.2.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.47.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.25.9
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==4.0.9
wrapt==1.12.1
zipp==3.1.0
